Update: Refactor API functions to another place
